requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
defusedxml==0.7.1
Deprecated==1.2.14
idna==3.10
jira==3.8.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
